requirements.txt
accelerate==0.16.0
albumentations==1.1.0
alembic==1.9.4
anyio @ file:///C:/ci/anyio_1644463701441/work/dist
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///C:/ci/argon2-cffi-bindings_1644569878360/work
astroid==2.9.3
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
autopep8==1.6.0
Babel @ file:///tmp/build/80754af9/babel_1620871417480/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
braceexpand==0.1.7
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
clip-anytorch==2.5.0
cmaes==0.9.1
CoCa-pytorch==0.0.7
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
colorlog==6.7.0
cryptography @ file:///C:/ci/cryptography_1639472366776/work
cycler==0.11.0
dalle2-pytorch==1.12.1
debugpy @ file:///C:/ci/debugpy_1637073815078/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
einops==0.6.0
einops-exts==0.0.4
ema-pytorch==0.2.1
embedding-reader==1.5.0
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.9.0
flake8==4.0.1
fonttools==4.29.1
fsspec==2023.1.0
ftfy==6.1.1
grad-cam==1.3.7
graphviz==0.20.1
greenlet==2.0.2
huggingface-hub==0.12.1
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.16.1
importlib-metadata @ file:///C:/ci/importlib-metadata_1638543108096/work
importlib-resources==5.12.0
ipykernel @ file:///C:/ci/ipykernel_1633545574250/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1643818321468/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
isort==5.10.1
jedi @ file:///C:/ci/jedi_1644315425835/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1635780242639/work
joblib==1.1.0
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1643638337975/work
jupyter-core @ file:///C:/ci/jupyter_core_1636537213213/work
jupyter-lsp==1.5.1
jupyter-server @ file:///opt/conda/conda-bld/jupyter_server_1644494914632/work
jupyterlab @ file:///opt/conda/conda-bld/jupyterlab_1644830542042/work
jupyterlab-lsp==3.10.1
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-server @ file:///opt/conda/conda-bld/jupyterlab_server_1644500396812/work
jupyterthemes==0.20.0
kiwisolver==1.3.2
kornia==0.6.10
lazy-object-proxy==1.7.1
lesscpy==0.15.1
llvmlite==0.40.1
lpips==0.1.4
Mako==1.2.4
MarkupSafe @ file:///C:/ci/markupsafe_1621528314575/work
matplotlib==3.5.1
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mccabe==0.6.1
mistune==0.8.4
nbclassic @ file:///opt/conda/conda-bld/nbclassic_1644943264176/work
nbclient @ file:///tmp/build/80754af9/nbclient_1645431659072/work
nbconvert @ file:///C:/ci/nbconvert_1624479163777/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx==2.7.1
notebook @ file:///C:/ci/notebook_1645002756862/work
numba==0.57.1
numpy==1.22.2
open-clip-torch==2.15.0
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.62
openpyxl==3.0.9
openslide-python==1.2.0
optuna==3.1.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.1
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
ply==3.11
prometheus-client @ file:///opt/conda/conda-bld/prometheus_client_1643788673601/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.20.3
psutil==5.9.4
pyarrow==7.0.0
pycodestyle==2.8.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.5
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pylint==2.12.2
pynndescent==0.5.10
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyrsistent @ file:///C:/ci/pyrsistent_1636111468851/work
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.4.0
pytorch-warmup==0.1.1
pytz==2021.3
PyWavelets==1.2.0
pywin32==302
pywinpty==1.1.6
PyYAML==6.0
pyzmq @ file:///C:/ci/pyzmq_1638435185959/work
qudida==0.0.4
regex==2022.10.31
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
resize-right==0.0.2
rope==0.23.0
rotary-embedding-torch==0.2.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
sentencepiece==0.1.97
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///C:/ci/sniffio_1614030707456/work
snowballstemmer==2.2.0
SQLAlchemy==2.0.4
terminado @ file:///C:/ci/terminado_1644322757089/work
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
threadpoolctl==3.1.0
tifffile==2022.2.9
timm==0.6.7
toml==0.10.2
torch==1.13.1+cu116
torch-fidelity==0.3.0
torchaudio==0.13.1
torchinfo==1.6.5
torchmetrics==0.11.3
torchvision==0.14.1+cu116
torchviz==0.0.2
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm==4.64.0
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
ttach==0.0.3
typing_extensions==4.5.0
ujson==5.1.0
umap-learn==0.5.3
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
vector-quantize-pytorch==1.0.7
vit-pytorch==0.35.8
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webdataset==0.2.33
webencodings==0.5.1
websocket-client @ file:///C:/ci/websocket-client_1614804473297/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
wrapt==1.13.3
x-clip==0.12.1
yapf==0.32.0
zipp @ file:///opt/conda/conda-bld/zipp_1641824620731/work
